python-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
numpy==2.2.2
faiss-cpu==1.10.0
sentence-transformers==3.3.1
rank-bm25==0.2.2
pydantic-settings==2.7.1
python-dotenv==1.0.1
langgraph==1.0.5
langchain-core==1.2.3
psycopg[binary]==3.2.12
httpx==0.28.1
pytest==8.3.4
